+++ src/shared/bodies/base_body.h
@@ -42,7 +42,7 @@
 #include "base_particle_generator.h"
 #include "base_particles.h"
 #include "cell_linked_list.h"
-#include "execution.h"
+#include "base_implementation.h"
 #include "sph_system.h"
 #include "sphinxsys_containers.h"
