+++ src/shared/particle_dynamics/particle_iterators.h
@@ -30,7 +30,7 @@
 #define PARTICLE_ITERATORS_H
 
 #include "base_data_package.h"
-#include "execution.h"
+#include "implementation.h"
 #include "sphinxsys_containers.h"
 
 #include <numeric>
