+++ src/shared/shared_ck/body_relation/relation_ck.h
@@ -31,7 +31,7 @@
 
 #include "base_body.h"
 #include "base_particles.h"
-#include "execution.h"
+#include "implementation.h"
 
 namespace SPH
 {
