+++ src/shared/shared_ck/particle_dynamics/loop_range.h
@@ -37,7 +37,7 @@ namespace SPH
 template <typename...>
 class LoopRangeCK;
 
-template <class ExecutionPolicy, class SPHBody>
+template <class ExecutionPolicy>
 class LoopRangeCK<ExecutionPolicy, SPHBody>
 {
     BaseParticles &particles_;
