+++ src/shared/shared_ck/particle_dynamics/particle_iterators_ck.h
@@ -29,7 +29,7 @@
 #ifndef PARTICLE_ITERATORS_CK_H
 #define PARTICLE_ITERATORS_CK_H
 
-#include "execution.h"
+#include "implementation.h"
 #include "loop_range.h"
 
 namespace SPH
